app_1/requirements.txt
google-adk==1.7.0
fastapi
uvicorn[standard]
google-cloud-aiplatform
python-multipart
google-cloud-firestore